backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = {extras = ["standart"], version = "^0.23.1"}
websockets = "^11.0.3"
pydantic-settings = "^2.0.2"
pytz = "^2023.3"
requests = "^2.31.0"
typer = {extras = ["all"], version = "^0.9.0"}
sqlalchemy = "^2.0.19"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
